Ajustes feitos nas entidades.

+++ dev/back-end/BackLaizeTech/BackLaizeTech/src/main/kotlin/laize_tech/back/ControllerJpa/CompraProdutoJpaController.kt
@@ -35,7 +35,7 @@ class CompraProdutoJpaController(
 //        }
 
         val produto = compraProdutoDTO.idProduto.let {
-            produtoRepository.findById(it.toLong()).orElseThrow {
+            produtoRepository.findById(it).orElseThrow {
                 IdNaoEncontradoException("Produto", it)
             }
         }
