+++ dev/back-end/BackLaizeTech/BackLaizeTech/src/main/kotlin/laize_tech/back/dto/ProdutoDTO.kt
@@ -1,7 +1,7 @@
 package laize_tech.back.dto
 
 data class ProdutoDTO(
-    val idCategoria: Long,
+    val idCategoria: Int,
     val nomeProduto: String,
     val quantidadeProduto: Int,
     val statusAtivo: Boolean,
